swiglu: move kernel header to experimental/ to match codegen convention
Quasar codegen-generated SFPU kernels live under
tt_llk_quasar/common/inc/experimental/ (currently abs and fill).
swiglu was placed under sfpu/ at PR creation; move it to experimental/
for consistency. Also update the include path in the test source.

No semantic change. Pytest collection unchanged: 131 tests.
